Extract bundling logic into its own module
