+++ .github/workflows/website-deploy.yml
@@ -37,6 +37,7 @@ jobs:
           -DBLAZE_DOCUMENTATION:BOOL=OFF
           -DBLAZE_EDITOR:BOOL=OFF
           -DBLAZE_FOUNDATION:BOOL=OFF
+          -DBLAZE_BUNDLE:BOOL=OFF
           -DBLAZE_TESTS:BOOL=OFF
           -DBLAZE_DOCS:BOOL=ON
       - run: cmake --build ./build --config Release --target doxygen
