+++ CMakeLists.txt
@@ -15,6 +15,7 @@ option(BLAZE_CODEGEN "Build the Blaze codegen library" ON)
 option(BLAZE_DOCUMENTATION "Build the Blaze documentation generator library" ON)
 option(BLAZE_EDITOR "Build the Blaze editor schema compatibility library" ON)
 option(BLAZE_FOUNDATION "Build the Blaze foundation JSON Schema library" ON)
+option(BLAZE_BUNDLE "Build the Blaze bundle library" ON)
 option(BLAZE_TESTS "Build the Blaze tests" OFF)
 option(BLAZE_BENCHMARK "Build the Blaze benchmarks" OFF)
 option(BLAZE_CONTRIB "Build the Blaze contrib programs" OFF)
@@ -49,6 +50,10 @@ endif()
 
 if(BLAZE_FOUNDATION)
   add_subdirectory(src/foundation)
+endif()
+
+if(BLAZE_BUNDLE)
+  add_subdirectory(src/bundle)
 endif()
 
 if(BLAZE_COMPILER)
@@ -135,6 +140,10 @@ if(BLAZE_TESTS)
 
   if(BLAZE_FOUNDATION)
     add_subdirectory(test/foundation)
+  endif()
+
+  if(BLAZE_BUNDLE)
+    add_subdirectory(test/bundle)
   endif()
 
   if(BLAZE_COMPILER)
