+++ config.cmake.in
@@ -14,6 +14,7 @@ if(NOT BLAZE_COMPONENTS)
   list(APPEND BLAZE_COMPONENTS codegen)
   list(APPEND BLAZE_COMPONENTS documentation)
   list(APPEND BLAZE_COMPONENTS editor)
+  list(APPEND BLAZE_COMPONENTS bundle)
 endif()
 
 include(CMakeFindDependencyMacro)
@@ -24,6 +25,7 @@ foreach(component ${BLAZE_COMPONENTS})
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
   elseif(component STREQUAL "compiler")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_bundle.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_evaluator.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_compiler.cmake")
   elseif(component STREQUAL "evaluator")
@@ -31,6 +33,7 @@ foreach(component ${BLAZE_COMPONENTS})
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_evaluator.cmake")
   elseif(component STREQUAL "test")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_bundle.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_evaluator.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_compiler.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_test.cmake")
@@ -40,6 +43,7 @@ foreach(component ${BLAZE_COMPONENTS})
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_output.cmake")
   elseif(component STREQUAL "configuration")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_bundle.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_configuration.cmake")
   elseif(component STREQUAL "alterschema")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
@@ -49,6 +53,7 @@ foreach(component ${BLAZE_COMPONENTS})
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_alterschema.cmake")
   elseif(component STREQUAL "codegen")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_bundle.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_evaluator.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_compiler.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_output.cmake")
@@ -60,6 +65,9 @@ foreach(component ${BLAZE_COMPONENTS})
   elseif(component STREQUAL "editor")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
     include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_editor.cmake")
+  elseif(component STREQUAL "bundle")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_foundation.cmake")
+    include("${CMAKE_CURRENT_LIST_DIR}/sourcemeta_blaze_bundle.cmake")
   else()
     message(FATAL_ERROR "Unknown Blaze component: ${component}")
   endif()
