+++ src/bundle/bundle.cc
@@ -1,3 +1,5 @@
+#include <sourcemeta/blaze/bundle.h>
+
 #include <sourcemeta/blaze/foundation.h>
 
 #include "helpers.h"
