+++ src/bundle/helpers.h
@@ -1,0 +1,86 @@
+#ifndef SOURCEMETA_BLAZE_BUNDLE_HELPERS_H
+#define SOURCEMETA_BLAZE_BUNDLE_HELPERS_H
+
+#include <sourcemeta/blaze/foundation.h>
+
+#include <cassert>     // assert
+#include <string_view> // std::string_view
+
+namespace sourcemeta::blaze {
+
+inline auto id_keyword(const SchemaBaseDialect base_dialect)
+    -> std::string_view {
+  switch (base_dialect) {
+    case SchemaBaseDialect::JSON_Schema_2020_12:
+    case SchemaBaseDialect::JSON_Schema_2020_12_Hyper:
+    case SchemaBaseDialect::JSON_Schema_2019_09:
+    case SchemaBaseDialect::JSON_Schema_2019_09_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_7:
+    case SchemaBaseDialect::JSON_Schema_Draft_7_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_6:
+    case SchemaBaseDialect::JSON_Schema_Draft_6_Hyper:
+      return "$id";
+    case SchemaBaseDialect::JSON_Schema_Draft_4:
+    case SchemaBaseDialect::JSON_Schema_Draft_4_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_3:
+    case SchemaBaseDialect::JSON_Schema_Draft_3_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_2_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_1_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_0_Hyper:
+      return "id";
+  }
+
+  assert(false);
+  return "$id";
+}
+
+inline auto definitions_keyword(const SchemaBaseDialect base_dialect)
+    -> std::string_view {
+  switch (base_dialect) {
+    case SchemaBaseDialect::JSON_Schema_2020_12:
+    case SchemaBaseDialect::JSON_Schema_2020_12_Hyper:
+    case SchemaBaseDialect::JSON_Schema_2019_09:
+    case SchemaBaseDialect::JSON_Schema_2019_09_Hyper:
+      return "$defs";
+    case SchemaBaseDialect::JSON_Schema_Draft_7:
+    case SchemaBaseDialect::JSON_Schema_Draft_7_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_6:
+    case SchemaBaseDialect::JSON_Schema_Draft_6_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_4:
+    case SchemaBaseDialect::JSON_Schema_Draft_4_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_3:
+    case SchemaBaseDialect::JSON_Schema_Draft_3_Hyper:
+      return "definitions";
+    case SchemaBaseDialect::JSON_Schema_Draft_2_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_1_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_0_Hyper:
+      return "";
+  }
+
+  assert(false);
+  return "$defs";
+}
+
+// In older drafts, the presence of `$ref` would override any sibling keywords
+// See
+// https://json-schema.org/draft-07/draft-handrews-json-schema-01#rfc.section.8.3
+inline auto
+ref_overrides_adjacent_keywords(const SchemaBaseDialect base_dialect) -> bool {
+  switch (base_dialect) {
+    case SchemaBaseDialect::JSON_Schema_Draft_7:
+    case SchemaBaseDialect::JSON_Schema_Draft_7_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_6:
+    case SchemaBaseDialect::JSON_Schema_Draft_6_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_4:
+    case SchemaBaseDialect::JSON_Schema_Draft_4_Hyper:
+    case SchemaBaseDialect::JSON_Schema_Draft_3:
+    case SchemaBaseDialect::JSON_Schema_Draft_3_Hyper:
+      return true;
+    default:
+      return false;
+  }
+}
+
+} // namespace sourcemeta::blaze
+
+#endif
