+++ src/bundle/include/sourcemeta/blaze/bundle.h
@@ -1,24 +1,31 @@
-#ifndef SOURCEMETA_BLAZE_FOUNDATION_BUNDLE_H
-#define SOURCEMETA_BLAZE_FOUNDATION_BUNDLE_H
+#ifndef SOURCEMETA_BLAZE_BUNDLE_H_
+#define SOURCEMETA_BLAZE_BUNDLE_H_
 
-#ifndef SOURCEMETA_BLAZE_FOUNDATION_EXPORT
-#include <sourcemeta/blaze/foundation_export.h>
+/// @defgroup bundle Bundle
+/// @brief Bundle JSON Schemas by inlining their external references.
+///
+/// This functionality is included as follows:
+///
+/// ```cpp
+/// #include <sourcemeta/blaze/bundle.h>
+/// ```
+
+#ifndef SOURCEMETA_BLAZE_BUNDLE_EXPORT
+#include <sourcemeta/blaze/bundle_export.h>
 #endif
 
 #include <sourcemeta/core/json.h>
 #include <sourcemeta/core/jsonpointer.h>
 
-// NOLINTBEGIN(misc-include-cleaner)
-#include <sourcemeta/blaze/foundation_frame.h>
-#include <sourcemeta/blaze/foundation_types.h>
-// NOLINTEND(misc-include-cleaner)
+#include <sourcemeta/blaze/foundation.h>
 
 #include <functional>  // std::function
+#include <optional>    // std::optional, std::nullopt
 #include <string_view> // std::string_view
 
 namespace sourcemeta::blaze {
 
-/// @ingroup foundation
+/// @ingroup bundle
 /// A callback to get dependency information
 /// - Origin URI (empty if none)
 /// - Pointer (reference keyword from the origin)
@@ -28,13 +35,14 @@ using DependencyCallback =
     std::function<void(std::string_view, const sourcemeta::core::WeakPointer &,
                        std::string_view, const sourcemeta::core::JSON &)>;
 
-/// @ingroup foundation
+/// @ingroup bundle
 ///
 /// This function recursively traverses and reports the external references in a
 /// schema. For example:
 ///
 /// ```cpp
 /// #include <sourcemeta/core/json.h>
+/// #include <sourcemeta/blaze/bundle.h>
 /// #include <sourcemeta/blaze/foundation.h>
 ///
 /// // A custom resolver that knows about an additional schema
@@ -66,7 +74,7 @@ using DependencyCallback =
 ///     // Do something with the information
 ///   });
 /// ```
-SOURCEMETA_BLAZE_FOUNDATION_EXPORT
+SOURCEMETA_BLAZE_BUNDLE_EXPORT
 auto dependencies(const sourcemeta::core::JSON &schema,
                   const SchemaWalker &walker, const SchemaResolver &resolver,
                   const DependencyCallback &callback,
@@ -75,7 +83,7 @@ auto dependencies(const sourcemeta::core::JSON &schema,
                   const SchemaFrame::Paths &paths = {
                       sourcemeta::core::empty_weak_pointer}) -> void;
 
-/// @ingroup foundation
+/// @ingroup bundle
 ///
 /// This function bundles a JSON Schema (starting from Draft 4) by embedding
 /// every remote reference into the top level schema resource, handling circular
@@ -83,6 +91,7 @@ auto dependencies(const sourcemeta::core::JSON &schema,
 ///
 /// ```cpp
 /// #include <sourcemeta/core/json.h>
+/// #include <sourcemeta/blaze/bundle.h>
 /// #include <sourcemeta/blaze/foundation.h>
 /// #include <cassert>
 ///
@@ -124,7 +133,7 @@ auto dependencies(const sourcemeta::core::JSON &schema,
 ///
 /// assert(document == expected);
 /// ```
-SOURCEMETA_BLAZE_FOUNDATION_EXPORT
+SOURCEMETA_BLAZE_BUNDLE_EXPORT
 auto bundle(sourcemeta::core::JSON &schema, const SchemaWalker &walker,
             const SchemaResolver &resolver,
             std::string_view default_dialect = "",
@@ -134,7 +143,7 @@ auto bundle(sourcemeta::core::JSON &schema, const SchemaWalker &walker,
             const SchemaFrame::Paths &paths = {
                 sourcemeta::core::empty_weak_pointer}) -> void;
 
-/// @ingroup foundation
+/// @ingroup bundle
 ///
 /// This function bundles a JSON Schema (starting from Draft 4) by embedding
 /// every remote reference into the top level schema resource, handling circular
@@ -143,6 +152,7 @@ auto bundle(sourcemeta::core::JSON &schema, const SchemaWalker &walker,
 ///
 /// ```cpp
 /// #include <sourcemeta/core/json.h>
+/// #include <sourcemeta/blaze/bundle.h>
 /// #include <sourcemeta/blaze/foundation.h>
 /// #include <cassert>
 ///
@@ -185,7 +195,7 @@ auto bundle(sourcemeta::core::JSON &schema, const SchemaWalker &walker,
 ///
 /// assert(result == expected);
 /// ```
-SOURCEMETA_BLAZE_FOUNDATION_EXPORT
+SOURCEMETA_BLAZE_BUNDLE_EXPORT
 auto bundle(
     const sourcemeta::core::JSON &schema, const SchemaWalker &walker,
     const SchemaResolver &resolver, std::string_view default_dialect = "",
