+++ src/compiler/CMakeLists.txt
@@ -26,5 +26,7 @@ target_link_libraries(sourcemeta_blaze_compiler PRIVATE
   sourcemeta::core::uri)
 target_link_libraries(sourcemeta_blaze_compiler PUBLIC
   sourcemeta::blaze::foundation)
+target_link_libraries(sourcemeta_blaze_compiler PRIVATE
+  sourcemeta::blaze::bundle)
 target_link_libraries(sourcemeta_blaze_compiler PUBLIC
   sourcemeta::blaze::evaluator)
