+++ src/compiler/compile.cc
@@ -1,6 +1,6 @@
+#include <sourcemeta/blaze/bundle.h>
 #include <sourcemeta/blaze/compiler.h>
 #include <sourcemeta/blaze/evaluator.h>
-
 #include <sourcemeta/blaze/foundation.h>
 
 #include <algorithm> // std::move, std::sort, std::unique
