+++ src/configuration/CMakeLists.txt
@@ -11,3 +11,4 @@ target_link_libraries(sourcemeta_blaze_configuration PRIVATE sourcemeta::core::u
 target_link_libraries(sourcemeta_blaze_configuration PRIVATE sourcemeta::core::io)
 target_link_libraries(sourcemeta_blaze_configuration PRIVATE sourcemeta::core::crypto)
 target_link_libraries(sourcemeta_blaze_configuration PRIVATE sourcemeta::blaze::foundation)
+target_link_libraries(sourcemeta_blaze_configuration PRIVATE sourcemeta::blaze::bundle)
