+++ test/bundle/bundle_draft2_test.cc
@@ -1,6 +1,8 @@
 #include <gtest/gtest.h>
 
+#include <sourcemeta/blaze/bundle.h>
 #include <sourcemeta/blaze/foundation.h>
+
 #include <sourcemeta/core/json.h>
 
 #include <string>      // std::string
@@ -10,7 +12,7 @@ static auto test_resolver(std::string_view identifier)
     -> std::optional<sourcemeta::core::JSON> {
   if (identifier == "https://www.sourcemeta.com/test-1") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-00/schema#",
+      "$schema": "http://json-schema.org/draft-02/schema#",
       "id": "https://www.sourcemeta.com/test-1",
       "type": "string"
     })JSON");
@@ -19,40 +21,40 @@ static auto test_resolver(std::string_view identifier)
   }
 }
 
-TEST(Foundation_bundle_draft0, no_references_no_id) {
+TEST(Bundle_draft2, no_references_no_id) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-00/schema#"
+    "$schema": "http://json-schema.org/draft-02/schema#"
   })JSON");
 
   sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
                             test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-00/schema#"
+    "$schema": "http://json-schema.org/draft-02/schema#"
   })JSON");
 
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft0, const_no_references_no_id) {
+TEST(Bundle_draft2, const_no_references_no_id) {
   const sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-00/schema#"
+    "$schema": "http://json-schema.org/draft-02/schema#"
   })JSON");
 
   const auto result = sourcemeta::blaze::bundle(
       document, sourcemeta::blaze::schema_walker, test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-00/schema#"
+    "$schema": "http://json-schema.org/draft-02/schema#"
   })JSON");
 
   EXPECT_EQ(result, expected);
 }
 
-TEST(Foundation_bundle_draft0, simple_bundling) {
+TEST(Bundle_draft2, simple_bundling) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-00/schema#",
+    "$schema": "http://json-schema.org/draft-02/schema#",
     "properties": {
       "test": { "$ref": "https://www.sourcemeta.com/test-1" }
     }
