+++ test/bundle/bundle_draft6_test.cc
@@ -1,6 +1,8 @@
 #include <gtest/gtest.h>
 
+#include <sourcemeta/blaze/bundle.h>
 #include <sourcemeta/blaze/foundation.h>
+
 #include <sourcemeta/core/json.h>
 
 #include <string>      // std::string
@@ -10,31 +12,31 @@ static auto test_resolver(std::string_view identifier)
     -> std::optional<sourcemeta::core::JSON> {
   if (identifier == "https://www.sourcemeta.com/test-1") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://www.sourcemeta.com/test-1",
       "type": "string"
     })JSON");
   } else if (identifier == "https://www.sourcemeta.com/test-2") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://www.sourcemeta.com/test-2",
       "allOf": [ { "$ref": "test-3" } ]
     })JSON");
   } else if (identifier == "https://www.sourcemeta.com/test-3") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://www.sourcemeta.com/test-3",
       "allOf": [ { "$ref": "test-1" } ]
     })JSON");
   } else if (identifier == "https://www.sourcemeta.com/test-4") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://www.sourcemeta.com/test-4",
       "type": "boolean"
     })JSON");
   } else if (identifier == "https://www.sourcemeta.com/recursive") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://www.sourcemeta.com/recursive",
       "properties": {
         "foo": { "$ref": "#" }
@@ -43,7 +45,7 @@ static auto test_resolver(std::string_view identifier)
   } else if (identifier ==
              "https://www.sourcemeta.com/recursive-empty-fragment") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://www.sourcemeta.com/recursive-empty-fragment#",
       "properties": {
         "foo": { "$ref": "#" }
@@ -55,7 +57,7 @@ static auto test_resolver(std::string_view identifier)
     })JSON");
   } else if (identifier == "https://www.sourcemeta.com/trailing-hash") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://www.sourcemeta.com/trailing-hash#",
       "definitions": {
         "string": { "type": "string" }
@@ -68,87 +70,48 @@ static auto test_resolver(std::string_view identifier)
     })JSON");
   } else if (identifier == "https://example.com/meta/2.json") {
     return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
+      "$schema": "http://json-schema.org/draft-06/schema#",
       "$id": "https://example.com/meta/2.json"
-    })JSON");
-  } else if (identifier == "https://example.com/draft7-dedup-a") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
-      "$id": "https://example.com/draft7-dedup-a",
-      "allOf": [ { "$ref": "draft7-dedup-common" } ],
-      "definitions": {
-        "https://example.com/draft7-dedup-common": {
-          "$schema": "http://json-schema.org/draft-07/schema#",
-          "$id": "https://example.com/draft7-dedup-common",
-          "type": "string"
-        }
-      }
-    })JSON");
-  } else if (identifier == "https://example.com/draft7-dedup-b") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
-      "$id": "https://example.com/draft7-dedup-b",
-      "allOf": [ { "$ref": "draft7-dedup-common" } ],
-      "definitions": {
-        "https://example.com/draft7-dedup-common": {
-          "$schema": "http://json-schema.org/draft-07/schema#",
-          "$id": "https://example.com/draft7-dedup-common",
-          "type": "string"
-        }
-      }
-    })JSON");
-  } else if (identifier == "https://example.com/draft7-elevate-single") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "http://json-schema.org/draft-07/schema#",
-      "$id": "https://example.com/draft7-elevate-single",
-      "allOf": [ { "$ref": "draft7-elevate-nested" } ],
-      "definitions": {
-        "https://example.com/draft7-elevate-nested": {
-          "$schema": "http://json-schema.org/draft-07/schema#",
-          "$id": "https://example.com/draft7-elevate-nested",
-          "type": "string"
-        }
-      }
     })JSON");
   } else {
     return sourcemeta::blaze::schema_resolver(identifier);
   }
 }
 
-TEST(Foundation_bundle_draft7, no_references_no_id) {
+TEST(Bundle_draft6, no_references_no_id) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#"
+    "$schema": "http://json-schema.org/draft-06/schema#"
   })JSON");
 
   sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
                             test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#"
+    "$schema": "http://json-schema.org/draft-06/schema#"
   })JSON");
 
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, const_no_references_no_id) {
+TEST(Bundle_draft6, const_no_references_no_id) {
   const sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#"
+    "$schema": "http://json-schema.org/draft-06/schema#"
   })JSON");
 
   const auto result = sourcemeta::blaze::bundle(
       document, sourcemeta::blaze::schema_walker, test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#"
+    "$schema": "http://json-schema.org/draft-06/schema#"
   })JSON");
 
   EXPECT_EQ(result, expected);
 }
 
-TEST(Foundation_bundle_draft7, simple_with_id) {
+TEST(Bundle_draft6, simple_with_id) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
       "bar": {
@@ -163,7 +126,7 @@ TEST(Foundation_bundle_draft7, simple_with_id) {
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
       "bar": {
@@ -173,17 +136,17 @@ TEST(Foundation_bundle_draft7, simple_with_id) {
     },
     "definitions": {
       "https://www.sourcemeta.com/test-1": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-1",
         "type": "string"
       },
       "https://www.sourcemeta.com/test-2": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-2",
         "allOf": [ { "$ref": "test-3" } ]
       },
       "https://www.sourcemeta.com/test-3": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-3",
         "allOf": [ { "$ref": "test-1" } ]
       }
@@ -193,9 +156,9 @@ TEST(Foundation_bundle_draft7, simple_with_id) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, simple_without_id) {
+TEST(Bundle_draft6, simple_without_id) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
       "bar": {
@@ -209,7 +172,7 @@ TEST(Foundation_bundle_draft7, simple_without_id) {
                             test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
       "bar": {
@@ -219,17 +182,17 @@ TEST(Foundation_bundle_draft7, simple_without_id) {
     },
     "definitions": {
       "https://www.sourcemeta.com/test-1": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-1",
         "type": "string"
       },
       "https://www.sourcemeta.com/test-2": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-2",
         "allOf": [ { "$ref": "test-3" } ]
       },
       "https://www.sourcemeta.com/test-3": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-3",
         "allOf": [ { "$ref": "test-1" } ]
       }
@@ -239,10 +202,10 @@ TEST(Foundation_bundle_draft7, simple_without_id) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, schema_not_found) {
+TEST(Bundle_draft6, schema_not_found) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "foo": { "$ref": "https://www.sourcemeta.com/xxx" }
     }
@@ -253,10 +216,10 @@ TEST(Foundation_bundle_draft7, schema_not_found) {
                sourcemeta::blaze::SchemaResolutionError);
 }
 
-TEST(Foundation_bundle_draft7, idempotency) {
+TEST(Bundle_draft6, idempotency) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "test": { "$ref": "https://www.sourcemeta.com/test-2" }
     }
@@ -271,23 +234,23 @@ TEST(Foundation_bundle_draft7, idempotency) {
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "test": { "$ref": "https://www.sourcemeta.com/test-2" }
     },
     "definitions": {
       "https://www.sourcemeta.com/test-1": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-1",
         "type": "string"
       },
       "https://www.sourcemeta.com/test-2": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-2",
         "allOf": [ { "$ref": "test-3" } ]
       },
       "https://www.sourcemeta.com/test-3": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-3",
         "allOf": [ { "$ref": "test-1" } ]
       }
@@ -297,16 +260,16 @@ TEST(Foundation_bundle_draft7, idempotency) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, pre_embedded) {
+TEST(Bundle_draft6, pre_embedded) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "test": { "$ref": "https://www.sourcemeta.com/test-2" }
     },
     "definitions": {
       "already-embedded": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-1",
         "type": "string"
       }
@@ -318,23 +281,23 @@ TEST(Foundation_bundle_draft7, pre_embedded) {
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "test": { "$ref": "https://www.sourcemeta.com/test-2" }
     },
     "definitions": {
       "already-embedded": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-1",
         "type": "string"
       },
       "https://www.sourcemeta.com/test-2": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-2",
         "allOf": [ { "$ref": "test-3" } ]
       },
       "https://www.sourcemeta.com/test-3": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-3",
         "allOf": [ { "$ref": "test-1" } ]
       }
@@ -344,10 +307,10 @@ TEST(Foundation_bundle_draft7, pre_embedded) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, taken_definitions_entry) {
+TEST(Bundle_draft6, taken_definitions_entry) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "test": { "$ref": "https://www.sourcemeta.com/test-1" },
       "extra": { "$ref": "https://www.sourcemeta.com/test-4" }
@@ -365,7 +328,7 @@ TEST(Foundation_bundle_draft7, taken_definitions_entry) {
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "test": { "$ref": "https://www.sourcemeta.com/test-1" },
       "extra": { "$ref": "https://www.sourcemeta.com/test-4" }
@@ -373,7 +336,7 @@ TEST(Foundation_bundle_draft7, taken_definitions_entry) {
     "definitions": {
       "https://www.sourcemeta.com/test-1": { "foo": 1 },
       "https://www.sourcemeta.com/test-1/x": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-1",
         "type": "string"
       },
@@ -381,7 +344,7 @@ TEST(Foundation_bundle_draft7, taken_definitions_entry) {
       "https://www.sourcemeta.com/test-4/x": { "foo": 1 },
       "https://www.sourcemeta.com/test-4/x/x": { "foo": 1 },
       "https://www.sourcemeta.com/test-4/x/x/x": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-4",
         "type": "boolean"
       }
@@ -391,9 +354,9 @@ TEST(Foundation_bundle_draft7, taken_definitions_entry) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, recursive) {
+TEST(Bundle_draft6, recursive) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "allOf": [ { "$ref": "https://www.sourcemeta.com/recursive" } ]
   })JSON");
 
@@ -401,11 +364,11 @@ TEST(Foundation_bundle_draft7, recursive) {
                             test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "allOf": [ { "$ref": "https://www.sourcemeta.com/recursive" } ],
     "definitions": {
       "https://www.sourcemeta.com/recursive": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/recursive",
         "properties": {
           "foo": { "$ref": "#" }
@@ -417,9 +380,9 @@ TEST(Foundation_bundle_draft7, recursive) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, recursive_empty_fragment) {
+TEST(Bundle_draft6, recursive_empty_fragment) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "allOf": [ { "$ref": "https://www.sourcemeta.com/recursive-empty-fragment#" } ]
   })JSON");
 
@@ -427,11 +390,11 @@ TEST(Foundation_bundle_draft7, recursive_empty_fragment) {
                             test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "allOf": [ { "$ref": "https://www.sourcemeta.com/recursive-empty-fragment#" } ],
     "definitions": {
       "https://www.sourcemeta.com/recursive-empty-fragment#": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/recursive-empty-fragment#",
         "properties": {
           "foo": { "$ref": "#" }
@@ -443,14 +406,14 @@ TEST(Foundation_bundle_draft7, recursive_empty_fragment) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, anonymous_no_dialect) {
+TEST(Bundle_draft6, anonymous_no_dialect) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "allOf": [ { "$ref": "https://www.sourcemeta.com/anonymous" } ]
   })JSON");
 
   sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
                             test_resolver,
-                            "http://json-schema.org/draft-07/schema#");
+                            "http://json-schema.org/draft-06/schema#");
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
     "allOf": [ { "$ref": "https://www.sourcemeta.com/anonymous" } ],
@@ -465,7 +428,7 @@ TEST(Foundation_bundle_draft7, anonymous_no_dialect) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, metaschema) {
+TEST(Bundle_draft6, metaschema) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$schema": "https://example.com/meta/1.json",
     "type": "string"
@@ -483,7 +446,7 @@ TEST(Foundation_bundle_draft7, metaschema) {
         "$id": "https://example.com/meta/1.json"
       },
       "https://example.com/meta/2.json": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://example.com/meta/2.json"
       }
     }
@@ -492,9 +455,9 @@ TEST(Foundation_bundle_draft7, metaschema) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, relative_base_uri_with_ref) {
+TEST(Bundle_draft6, relative_base_uri_with_ref) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "$id": "common",
     "allOf": [ { "$ref": "#reference" } ],
     "definitions": {
@@ -508,7 +471,7 @@ TEST(Foundation_bundle_draft7, relative_base_uri_with_ref) {
                             test_resolver);
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "$id": "common",
     "allOf": [ { "$ref": "#reference" } ],
     "definitions": {
@@ -521,11 +484,11 @@ TEST(Foundation_bundle_draft7, relative_base_uri_with_ref) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, hyperschema_smoke) {
+TEST(Bundle_draft6, hyperschema_smoke) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/hyper-schema#",
+    "$schema": "http://json-schema.org/draft-06/hyper-schema#",
     "allOf": [
-      { "$ref": "http://json-schema.org/draft-07/hyper-schema#" }
+      { "$ref": "http://json-schema.org/draft-06/hyper-schema#" }
     ]
   })JSON");
 
@@ -535,12 +498,12 @@ TEST(Foundation_bundle_draft7, hyperschema_smoke) {
   EXPECT_TRUE(document.is_object());
 }
 
-TEST(Foundation_bundle_draft7, hyperschema_1) {
+TEST(Bundle_draft6, hyperschema_1) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "allOf": [
-      { "$ref": "http://json-schema.org/draft-07/schema#" },
-      { "$ref": "http://json-schema.org/draft-07/hyper-schema#" }
+      { "$ref": "http://json-schema.org/draft-06/schema#" },
+      { "$ref": "http://json-schema.org/draft-06/hyper-schema#" }
     ]
   })JSON");
 
@@ -549,21 +512,19 @@ TEST(Foundation_bundle_draft7, hyperschema_1) {
 
   EXPECT_TRUE(document.defines("definitions"));
   EXPECT_TRUE(document.at("definitions").is_object());
-  EXPECT_EQ(document.at("definitions").size(), 3);
+  EXPECT_EQ(document.at("definitions").size(), 2);
 
   EXPECT_TRUE(document.at("definitions")
-                  .defines("http://json-schema.org/draft-07/schema#"));
+                  .defines("http://json-schema.org/draft-06/schema#"));
   EXPECT_TRUE(document.at("definitions")
-                  .defines("http://json-schema.org/draft-07/links#"));
-  EXPECT_TRUE(document.at("definitions")
-                  .defines("http://json-schema.org/draft-07/hyper-schema#"));
+                  .defines("http://json-schema.org/draft-06/hyper-schema#"));
 }
 
-TEST(Foundation_bundle_draft7, hyperschema_ref_metaschema) {
+TEST(Bundle_draft6, hyperschema_ref_metaschema) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/hyper-schema#",
+    "$schema": "http://json-schema.org/draft-06/hyper-schema#",
     "allOf": [
-      { "$ref": "http://json-schema.org/draft-07/schema#" }
+      { "$ref": "http://json-schema.org/draft-06/schema#" }
     ]
   })JSON");
 
@@ -575,50 +536,23 @@ TEST(Foundation_bundle_draft7, hyperschema_ref_metaschema) {
   EXPECT_EQ(document.at("definitions").size(), 1);
 
   EXPECT_TRUE(document.at("definitions")
-                  .defines("http://json-schema.org/draft-07/schema#"));
+                  .defines("http://json-schema.org/draft-06/schema#"));
 }
 
-TEST(Foundation_bundle_draft7, bundle_to_defs) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$ref": "https://www.sourcemeta.com/recursive"
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver, "", "",
-                            sourcemeta::core::Pointer{"$defs"});
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$ref": "https://www.sourcemeta.com/recursive",
-    "$defs": {
-      "https://www.sourcemeta.com/recursive": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "$id": "https://www.sourcemeta.com/recursive",
-        "properties": {
-          "foo": { "$ref": "#" }
-        }
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_draft7, standalone_ref_with_default_dialect) {
+TEST(Bundle_draft6, standalone_ref_with_default_dialect) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$ref": "https://www.sourcemeta.com/test-1"
   })JSON");
 
   sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
                             test_resolver,
-                            "http://json-schema.org/draft-07/schema#");
+                            "http://json-schema.org/draft-06/schema#");
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
     "allOf": [ { "$ref": "https://www.sourcemeta.com/test-1" } ],
     "definitions": {
       "https://www.sourcemeta.com/test-1": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/test-1",
         "type": "string"
       }
@@ -628,10 +562,10 @@ TEST(Foundation_bundle_draft7, standalone_ref_with_default_dialect) {
   EXPECT_EQ(document, expected);
 }
 
-TEST(Foundation_bundle_draft7, ref_with_fragment_to_id_with_trailing_hash) {
+TEST(Bundle_draft6, ref_with_fragment_to_id_with_trailing_hash) {
   sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "foo": {
         "$ref": "https://www.sourcemeta.com/trailing-hash#/definitions/string"
@@ -644,7 +578,7 @@ TEST(Foundation_bundle_draft7, ref_with_fragment_to_id_with_trailing_hash) {
 
   const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
     "$id": "https://example.com",
-    "$schema": "http://json-schema.org/draft-07/schema#",
+    "$schema": "http://json-schema.org/draft-06/schema#",
     "properties": {
       "foo": {
         "$ref": "https://www.sourcemeta.com/trailing-hash#/definitions/string"
@@ -652,84 +586,11 @@ TEST(Foundation_bundle_draft7, ref_with_fragment_to_id_with_trailing_hash) {
     },
     "definitions": {
       "https://www.sourcemeta.com/trailing-hash#": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
+        "$schema": "http://json-schema.org/draft-06/schema#",
         "$id": "https://www.sourcemeta.com/trailing-hash#",
         "definitions": {
           "string": { "type": "string" }
         }
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_draft7, deduplicate_embedded_from_prebundled) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://example.com/draft7-dedup-entry",
-    "allOf": [
-      { "$ref": "draft7-dedup-a" },
-      { "$ref": "draft7-dedup-b" }
-    ]
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://example.com/draft7-dedup-entry",
-    "allOf": [
-      { "$ref": "draft7-dedup-a" },
-      { "$ref": "draft7-dedup-b" }
-    ],
-    "definitions": {
-      "https://example.com/draft7-dedup-a": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "$id": "https://example.com/draft7-dedup-a",
-        "allOf": [ { "$ref": "draft7-dedup-common" } ]
-      },
-      "https://example.com/draft7-dedup-b": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "$id": "https://example.com/draft7-dedup-b",
-        "allOf": [ { "$ref": "draft7-dedup-common" } ]
-      },
-      "https://example.com/draft7-dedup-common": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "$id": "https://example.com/draft7-dedup-common",
-        "type": "string"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_draft7, elevate_embedded_from_single_prebundled) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://example.com/draft7-elevate-entry",
-    "allOf": [ { "$ref": "draft7-elevate-single" } ]
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "http://json-schema.org/draft-07/schema#",
-    "$id": "https://example.com/draft7-elevate-entry",
-    "allOf": [ { "$ref": "draft7-elevate-single" } ],
-    "definitions": {
-      "https://example.com/draft7-elevate-single": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "$id": "https://example.com/draft7-elevate-single",
-        "allOf": [ { "$ref": "draft7-elevate-nested" } ]
-      },
-      "https://example.com/draft7-elevate-nested": {
-        "$schema": "http://json-schema.org/draft-07/schema#",
-        "$id": "https://example.com/draft7-elevate-nested",
-        "type": "string"
       }
     }
   })JSON");
