+++ test/editor/CMakeLists.txt
@@ -7,4 +7,6 @@ target_link_libraries(sourcemeta_blaze_editor_unit
 target_link_libraries(sourcemeta_blaze_editor_unit
   PRIVATE sourcemeta::blaze::foundation)
 target_link_libraries(sourcemeta_blaze_editor_unit
+  PRIVATE sourcemeta::blaze::bundle)
+target_link_libraries(sourcemeta_blaze_editor_unit
   PRIVATE sourcemeta::blaze::editor)
