+++ test/evaluator/CMakeLists.txt
@@ -17,6 +17,8 @@ target_link_libraries(sourcemeta_blaze_evaluator_unit
 target_link_libraries(sourcemeta_blaze_evaluator_unit
   PRIVATE sourcemeta::blaze::foundation)
 target_link_libraries(sourcemeta_blaze_evaluator_unit
+  PRIVATE sourcemeta::blaze::bundle)
+target_link_libraries(sourcemeta_blaze_evaluator_unit
   PRIVATE sourcemeta::blaze::compiler)
 target_link_libraries(sourcemeta_blaze_evaluator_unit
   PRIVATE sourcemeta::blaze::evaluator)
