--- test/foundation/foundation_bundle_2019_09_test.cc
@@ -1,566 +1,0 @@
-#include <gtest/gtest.h>
-
-#include <sourcemeta/blaze/foundation.h>
-#include <sourcemeta/core/json.h>
-
-#include <string>      // std::string
-#include <string_view> // std::string_view
-
-static auto test_resolver(std::string_view identifier)
-    -> std::optional<sourcemeta::core::JSON> {
-  if (identifier == "https://example.com/foo/bar") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://example.com/foo/bar",
-      "$anchor": "baz"
-    })JSON");
-  } else if (identifier == "https://example.com/baz-anchor") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://example.com/baz-anchor",
-      "$defs": {
-        "baz": {
-          "$anchor": "baz",
-          "type": "string"
-        }
-      }
-    })JSON");
-  } else if (identifier == "https://www.sourcemeta.com/test-1") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://www.sourcemeta.com/test-1",
-      "type": "string"
-    })JSON");
-  } else if (identifier == "https://www.sourcemeta.com/test-2") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://www.sourcemeta.com/test-2",
-      "$ref": "test-3"
-    })JSON");
-  } else if (identifier == "https://www.sourcemeta.com/test-3") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://www.sourcemeta.com/test-3",
-      "$ref": "test-1"
-    })JSON");
-  } else if (identifier == "https://www.sourcemeta.com/test-4") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://www.sourcemeta.com/test-4",
-      "type": "boolean"
-    })JSON");
-  } else if (identifier == "https://www.sourcemeta.com/recursive") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://www.sourcemeta.com/recursive",
-      "properties": {
-        "foo": { "$ref": "#" }
-      }
-    })JSON");
-  } else if (identifier ==
-             "https://www.sourcemeta.com/recursive-empty-fragment") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://www.sourcemeta.com/recursive-empty-fragment#",
-      "properties": {
-        "foo": { "$ref": "#" }
-      }
-    })JSON");
-  } else if (identifier == "https://www.sourcemeta.com/anonymous") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "type": "integer"
-    })JSON");
-  } else if (identifier == "https://example.com/meta/1.json") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://example.com/meta/2.json",
-      "$id": "https://example.com/meta/1.json",
-      "$vocabulary": { "https://json-schema.org/draft/2019-09/vocab/core": true }
-    })JSON");
-  } else if (identifier == "https://example.com/meta/2.json") {
-    return sourcemeta::core::parse_json(R"JSON({
-      "$schema": "https://json-schema.org/draft/2019-09/schema",
-      "$id": "https://example.com/meta/2.json",
-      "$vocabulary": { "https://json-schema.org/draft/2019-09/vocab/core": true }
-    })JSON");
-  } else {
-    return sourcemeta::blaze::schema_resolver(identifier);
-  }
-}
-
-TEST(Foundation_bundle_2019_09, no_references_no_id) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema"
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema"
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, const_no_references_no_id) {
-  const sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema"
-  })JSON");
-
-  const auto result = sourcemeta::blaze::bundle(
-      document, sourcemeta::blaze::schema_walker, test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema"
-  })JSON");
-
-  EXPECT_EQ(result, expected);
-}
-
-TEST(Foundation_bundle_2019_09, simple_with_id) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
-      "bar": { "$id": "https://www.sourcemeta.com", "$ref": "test-2" },
-      "baz": { "$ref": "https://example.com/foo/bar#baz" },
-      "qux": { "$ref": "https://example.com/foo/bar" }
-    }
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
-      "bar": { "$id": "https://www.sourcemeta.com", "$ref": "test-2" },
-      "baz": { "$ref": "https://example.com/foo/bar#baz" },
-      "qux": { "$ref": "https://example.com/foo/bar" }
-    },
-    "$defs": {
-      "https://www.sourcemeta.com/test-1": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-1",
-        "type": "string"
-      },
-      "https://www.sourcemeta.com/test-2": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-2",
-        "$ref": "test-3"
-      },
-      "https://www.sourcemeta.com/test-3": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-3",
-        "$ref": "test-1"
-      },
-      "https://example.com/foo/bar": {
-        "$id": "https://example.com/foo/bar",
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$anchor": "baz"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, simple_without_id) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
-      "bar": { "$id": "https://www.sourcemeta.com", "$ref": "test-2" },
-      "baz": { "$ref": "https://example.com/foo/bar#baz" },
-      "qux": { "$ref": "https://example.com/foo/bar" }
-    }
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "foo": { "$ref": "https://www.sourcemeta.com/test-1" },
-      "bar": { "$id": "https://www.sourcemeta.com", "$ref": "test-2" },
-      "baz": { "$ref": "https://example.com/foo/bar#baz" },
-      "qux": { "$ref": "https://example.com/foo/bar" }
-    },
-    "$defs": {
-      "https://www.sourcemeta.com/test-1": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-1",
-        "type": "string"
-      },
-      "https://www.sourcemeta.com/test-2": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-2",
-        "$ref": "test-3"
-      },
-      "https://www.sourcemeta.com/test-3": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-3",
-        "$ref": "test-1"
-      },
-      "https://example.com/foo/bar": {
-        "$id": "https://example.com/foo/bar",
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$anchor": "baz"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, schema_not_found) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "foo": { "$ref": "https://www.sourcemeta.com/xxx" }
-    }
-  })JSON");
-
-  EXPECT_THROW(sourcemeta::blaze::bundle(
-                   document, sourcemeta::blaze::schema_walker, test_resolver),
-               sourcemeta::blaze::SchemaResolutionError);
-}
-
-TEST(Foundation_bundle_2019_09, idempotency) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "test": { "$ref": "https://www.sourcemeta.com/test-2" }
-    }
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "test": { "$ref": "https://www.sourcemeta.com/test-2" }
-    },
-    "$defs": {
-      "https://www.sourcemeta.com/test-1": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-1",
-        "type": "string"
-      },
-      "https://www.sourcemeta.com/test-2": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-2",
-        "$ref": "test-3"
-      },
-      "https://www.sourcemeta.com/test-3": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-3",
-        "$ref": "test-1"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, pre_embedded) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "test": { "$ref": "https://www.sourcemeta.com/test-2" }
-    },
-    "$defs": {
-      "already-embedded": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-1",
-        "type": "string"
-      }
-    }
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "test": { "$ref": "https://www.sourcemeta.com/test-2" }
-    },
-    "$defs": {
-      "already-embedded": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-1",
-        "type": "string"
-      },
-      "https://www.sourcemeta.com/test-2": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-2",
-        "$ref": "test-3"
-      },
-      "https://www.sourcemeta.com/test-3": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-3",
-        "$ref": "test-1"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, taken_definitions_entry) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "test": { "$ref": "https://www.sourcemeta.com/test-1" },
-      "extra": { "$ref": "https://www.sourcemeta.com/test-4" }
-    },
-    "$defs": {
-      "https://www.sourcemeta.com/test-1": { "foo": 1 },
-      "https://www.sourcemeta.com/test-4": { "foo": 1 },
-      "https://www.sourcemeta.com/test-4/x": { "foo": 1 },
-      "https://www.sourcemeta.com/test-4/x/x": { "foo": 1 }
-    }
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$id": "https://example.com",
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "properties": {
-      "test": { "$ref": "https://www.sourcemeta.com/test-1" },
-      "extra": { "$ref": "https://www.sourcemeta.com/test-4" }
-    },
-    "$defs": {
-      "https://www.sourcemeta.com/test-1": { "foo": 1 },
-      "https://www.sourcemeta.com/test-1/x": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-1",
-        "type": "string"
-      },
-      "https://www.sourcemeta.com/test-4": { "foo": 1 },
-      "https://www.sourcemeta.com/test-4/x": { "foo": 1 },
-      "https://www.sourcemeta.com/test-4/x/x": { "foo": 1 },
-      "https://www.sourcemeta.com/test-4/x/x/x": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/test-4",
-        "type": "boolean"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, recursive) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "$ref": "https://www.sourcemeta.com/recursive"
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "$ref": "https://www.sourcemeta.com/recursive",
-    "$defs": {
-      "https://www.sourcemeta.com/recursive": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/recursive",
-        "properties": {
-          "foo": { "$ref": "#" }
-        }
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, recursive_empty_fragment) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "$ref": "https://www.sourcemeta.com/recursive-empty-fragment#"
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "$ref": "https://www.sourcemeta.com/recursive-empty-fragment#",
-    "$defs": {
-      "https://www.sourcemeta.com/recursive-empty-fragment#": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://www.sourcemeta.com/recursive-empty-fragment#",
-        "properties": {
-          "foo": { "$ref": "#" }
-        }
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, anonymous_no_dialect) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$ref": "https://www.sourcemeta.com/anonymous"
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver,
-                            "https://json-schema.org/draft/2019-09/schema");
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$ref": "https://www.sourcemeta.com/anonymous",
-    "$defs": {
-      "https://www.sourcemeta.com/anonymous": {
-        "$id": "https://www.sourcemeta.com/anonymous",
-        "type": "integer"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, metaschema) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://example.com/meta/1.json",
-    "type": "string"
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://example.com/meta/1.json",
-    "type": "string",
-    "$defs": {
-      "https://example.com/meta/1.json": {
-        "$schema": "https://example.com/meta/2.json",
-        "$id": "https://example.com/meta/1.json",
-        "$vocabulary": { "https://json-schema.org/draft/2019-09/vocab/core": true }
-      },
-      "https://example.com/meta/2.json": {
-        "$schema": "https://json-schema.org/draft/2019-09/schema",
-        "$id": "https://example.com/meta/2.json",
-        "$vocabulary": { "https://json-schema.org/draft/2019-09/vocab/core": true }
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, relative_base_uri_with_ref) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "$id": "common",
-    "allOf": [ { "$ref": "#reference" } ],
-    "definitions": {
-      "reference": {
-        "$anchor": "reference"
-      }
-    }
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  const sourcemeta::core::JSON expected = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "$id": "common",
-    "allOf": [ { "$ref": "#reference" } ],
-    "definitions": {
-      "reference": {
-        "$anchor": "reference"
-      }
-    }
-  })JSON");
-
-  EXPECT_EQ(document, expected);
-}
-
-TEST(Foundation_bundle_2019_09, hyperschema_smoke) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "$ref": "https://json-schema.org/draft/2019-09/hyper-schema"
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  EXPECT_TRUE(document.is_object());
-}
-
-TEST(Foundation_bundle_2019_09, hyperschema_1) {
-  sourcemeta::core::JSON document = sourcemeta::core::parse_json(R"JSON({
-    "$schema": "https://json-schema.org/draft/2019-09/schema",
-    "allOf": [
-      { "$ref": "https://json-schema.org/draft/2019-09/schema" },
-      { "$ref": "https://json-schema.org/draft/2019-09/meta/hyper-schema" }
-    ]
-  })JSON");
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  EXPECT_TRUE(document.defines("$defs"));
-  EXPECT_TRUE(document.at("$defs").is_object());
-  EXPECT_EQ(document.at("$defs").size(), 10);
-
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/schema"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/meta/core"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/meta/applicator"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/meta/validation"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/meta/meta-data"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/meta/format"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/meta/content"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/meta/hyper-schema"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/links"));
-  EXPECT_TRUE(document.at("$defs").defines(
-      "https://json-schema.org/draft/2019-09/hyper-schema"));
-}
-
-TEST(Foundation_bundle_2019_09, hyperschema_2) {
-  sourcemeta::core::JSON document =
-      sourcemeta::blaze::schema_resolver(
-          "https://json-schema.org/draft/2019-09/hyper-schema")
-          .value();
-
-  sourcemeta::blaze::bundle(document, sourcemeta::blaze::schema_walker,
-                            test_resolver);
-
-  EXPECT_TRUE(document.defines("$defs"));
-  EXPECT_TRUE(document.at("$defs").is_object());
-  EXPECT_EQ(document.at("$defs").size(), 9);
-}
