+++ test/packaging/find_package/CMakeLists.txt
@@ -16,3 +16,4 @@ target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::codegen)
 target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::documentation)
 target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::editor)
 target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::foundation)
+target_link_libraries(blaze_hello PRIVATE sourcemeta::blaze::bundle)
