+++ test/packaging/find_package/hello.cc
@@ -1,7 +1,5 @@
-#include <sourcemeta/blaze/foundation.h>
-#include <sourcemeta/core/json.h>
-
 #include <sourcemeta/blaze/alterschema.h>
+#include <sourcemeta/blaze/bundle.h>
 #include <sourcemeta/blaze/codegen.h>
 #include <sourcemeta/blaze/compiler.h>
 #include <sourcemeta/blaze/configuration.h>
@@ -11,6 +9,8 @@
 #include <sourcemeta/blaze/foundation.h>
 #include <sourcemeta/blaze/output.h>
 #include <sourcemeta/blaze/test.h>
+
+#include <sourcemeta/core/json.h>
 
 #include <cstdlib>  // EXIT_SUCCESS
 #include <iostream> // std::cerr
@@ -48,6 +48,13 @@ auto main() -> int {
       sourcemeta::blaze::SchemaFrame::Mode::Locations};
   foundation_frame.analyse(foundation_schema, sourcemeta::blaze::schema_walker,
                            sourcemeta::blaze::schema_resolver);
+
+  auto bundle_schema{sourcemeta::core::parse_json(R"JSON({
+    "$schema": "https://json-schema.org/draft/2020-12/schema",
+    "type": "string"
+  })JSON")};
+  sourcemeta::blaze::bundle(bundle_schema, sourcemeta::blaze::schema_walker,
+                            sourcemeta::blaze::schema_resolver);
 
   return EXIT_SUCCESS;
 }
